fix: consolidate contracttests files for pkg-file-count limit
Merge generated response-event tests into openapi_contract_response_events_test.go and fold shared-factory-starter coverage into openapi_contract_explicit_root_helper_test.go so pkg/api/contracttests stays at 15 files.
